pyproject.toml
[tool.poetry]
name = "b2broker-test-assignment"
version = "0.1.0"
description = ""
authors = ["Roman Lee"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
djangorestframework = "^3.14.0"
djangorestframework-jsonapi = {version = "^6.1.0", extras = ["django-filter"]}
django-filter = "^23.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
